type: avoid dynamic allocation inside the yaksi_type
Most most of the backend object to statically become a part of the
yaksi_type object.  This allows the backend pieces to be maintained as
a part of the pool too.  A shortcoming of this patch is that some of
the backend private content is exposed to the upper layer in the
header file.  But with appropriate naming, we can perhaps (virtually)
hide them.

Signed-off-by: Pavan Balaji <balaji@anl.gov>