cuda: make stream fully asynchronous
cudaStreamCreate creates a stream that implicitly synchronizes with the
default stream (stream 0). Thus, if work on the created stream is
interleaved with work on the default stream the two will be serialized.
This behavior can be fixed by replacing cudaStreamCreate with
cudaStreamCreateWithFlags and using the cudaStreamNonBlocking flag for
the new stream.

Signed-off-by: Giuseppe Congiu <gcongiu@anl.gov>