backend: improve hiding backend private content
A lot of the backend private content was included in the pre/post
headers, which are propagated upstream.  This patch pushes the private
content behind a void *, so layers that should not have access to that
content don't accidentally see them (the compiler should be able to
catch such accidental breakage).  This is basically the poor-man's
version of "class private" variables.

Signed-off-by: Pavan Balaji <balaji@anl.gov>