events: move event creation to the backend device layer
Device creation needs to take GPU locality into account.  For example,
with CUDA, the event creation depends on what device context we are
set to.  Moving the event creation to the device layer allows the
device to make the decision about creating the event when it is
actually needed.  At this time, it'd also have the information on
which device context the event needs to be created on.

Signed-off-by: Pavan Balaji <balaji@anl.gov>