backend/cuda: create per-device streams, events, and temporary buffers
The original code was using a single stream and temporary buffers
irrespective of the number of devices and would not work when the
number of GPUs available/used is more than one.  This patch makes
yaksa aware of the multiple GPUs and sets the right device context
before using streams and events.

Signed-off-by: Pavan Balaji <balaji@anl.gov>