backend: split generated content into smaller files
We split it into one file for the top two levels of the datatype.  A
shortcoming of this approach is that the top two levels are somewhat
"static" in that they are always compiled, i.e., they cannot be
controlled by the --pup-max-nesting argument.

Unfortunately, this patch includes a lot of code churn to move the
redundant code into a common module, instead of replicating the logic
between the seq and cuda backends.

Fixes pmodels/yaksa#21

Signed-off-by: Pavan Balaji <balaji@anl.gov>