unpack: return the actual number of unpacked bytes
When the user receives data in a pipelined fashion, it is possible
that the number of bytes received might not be able to fully fit a
basic datatype element.  In such cases, yaksa needs to tell the user
how much data was actually unpacked, so the user can keep track of the
remaining bytes.

Fixes pmodels/yaksa#67

Signed-off-by: Pavan Balaji <balaji@anl.gov>