types: optimize overgeneralized datatypes
Optimize for the case where the user overgeneralized a datatype, when
the datatype can really be represented as a simpler type.  This is
particularly important for struct because structs are not optimized
for pack/unpack, and they can create an arbitrarily large number of
child datatypes during unflattening.

There are several other optimizations possible, for example, for cases
where the majority of the types in the struct are of a single type.
In such cases, the struct could be split into an hindexed for the part
that uses a single type and a struct for the rest.  This patch does
not optimize such cases.  For now, we optimize the simple case where
all child types are the same.

Signed-off-by: Pavan Balaji <balaji@anl.gov>