warnings: avoid inline initialization of POD variables
As per the C++ standard:

"It is possible to transfer into a block, but not in a way that
bypasses declarations with initialization. A program that jumps from a
point where a local variable with automatic storage duration is not in
scope to a point where it is in scope is ill-formed unless the
variable has POD type (3.9) and is declared without an initializer"

I haven't searched for it in the C99 standard, but the concerns are
valid even for C99 and compilers such as PGI seem to throw warnings
for it even with C99 code.  Separating out the declaration statement
from the initialization statement should workaround these warnings.

Signed-off-by: Pavan Balaji <balaji@anl.gov>