dtpools: upgrade internal extent to 64 bits
We calculate the potential extent of a datatype, make sure it fits
inside yaksa's extent type format, and then decide whether we can use
this datatype or not.  For this, we need the internal calculation of
the extent to always be done with 64 bits, so we can eventually check
if it is within the required range or not.

Signed-off-by: Pavan Balaji <balaji@anl.gov>