cuda: remove reliance on yaksi for nvcc-compiled code
The NVCC compiled code needs to be fairly self contained and should
not rely on any of the configure-detected macros.  Configure uses the
primary compiler and not nvcc for detecting various features, which
might not match what nvcc provides.  In particular, nvcc might not
have C11 atomics, even if the primary compiler supports them.

Signed-off-by: Pavan Balaji <balaji@anl.gov>