backend/cuda: remove unnecessary checks for buffer types
Remove the unnecessary checks for the buffer type before doing the
pack/unpack.  These are already done at the glue layer before handing
it down to the driver layer.  This also avoids incorrect guesses by
the driver layer in case the glue layer passes down a temporary buffer
in place of the actual user buffer.

Signed-off-by: Pavan Balaji <balaji@anl.gov>