test/pack: fixes to the pack test
The pack test forces the user to specify the device IDs and the
memtypes, which are incorrect for multi-threaded tests.

The device ID increment was not thread safe, and would cause race
conditions in the test, thus causing the results to be
nondeterministic.  This patch creates a random set of device_ids at
the start of the program, so the runs are deterministic.

The memtypes are also generated deterministically randomly now, so we
don't need separate tests for each memtype.

Co-authored-by: Gengbin Zheng <gengbin.zheng@intel.com>

Signed-off-by: Pavan Balaji <balaji@anl.gov>