cuda: Allow configure to continue after nvcc check
If CUDA was not explicitly requested, it is possible that configure will
find the CUDA runtime library, then throw an error when it fails to find
nvcc. Remove AC_MSG_ERROR from the nvcc check, and instead let the
have_cuda/with_cuda variables determine if configure should continue
without CUDA support or throw an error. This fixes a build issue in the
Spack Gitlab CI environment.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>