backend: fix p2p check when mapped device is used
If a mapped device is supplied via info hint for d2d transfers, that
device needs to be used when checking for peer access. Otherwise, an
illegal memory could occur if peer access is not enabled from the mapped
device to the input buffer. This resolves a crash in MPICH after the
switch to lazily initialize CUDA resources.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>