handles: Remove unnecessary const qualifier
With GCC 11, the const qualifier for the data argument in
yaksu_handle_pool_elem_alloc causes the compiler to think memory might
be read before it is initialized (-Wmaybe-uninitialized). But the buffer
in question is never read in the function, so const most likely makes no
difference. Remove it and fix the warning.

Signed-off-by: Ken Raffenetti <raffenet@mcs.anl.gov>